Add files via upload
In this project, I'm going to practice using my SQL skills to answer business questions. This is a project extracted from the **`DataQuest: Data Engineer Track`**

I'll be using the Chinook database.  The Chinook database contains information about a fictional digital music shop - kind of like a mini-iTunes store. The Chinook database contains information about the artists, songs, and albums from the music shop, as well as information on the shop's employees, customers, and the customers purchases. This information is contained in eleven tables. The Chinook database is provided as a SQLite database file called chinook.db. 

I'll also be using a bit of pandas in this project to write queries that produce dataframe outputs.

I'll be posting the questions and answering them using my code